% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnsOutputs.R
\name{create_sparse_output}
\alias{create_sparse_output}
\title{Convert to sparse format}
\usage{
create_sparse_output(comparison_output, orig_group_cols = NULL)
}
\arguments{
\item{comparison_output}{Output from the comparison Table functions}

\item{orig_group_cols}{Vector of the same column names use used in group_col
when making the comparison_output object.}
}
\description{
Each row of the output sparse data frame has an individual value which has
been changed. This is useful for really wide data frames when only a few values
have changed. It can be easier in some cases to have a table of changes. The
output columns are: group.name, change_type, column, old_value, new_value
}
