% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnsOutputs.R
\name{create_wide_output}
\alias{create_wide_output}
\title{Convert to wide format}
\usage{
create_wide_output(comparison_output, suffix = c("_new", "_old"))
}
\arguments{
\item{comparison_output}{Output from the comparison Table functions}

\item{suffix}{Nomenclature for the new and old dataframe}
}
\description{
Easier to compare side-by-side
}
