% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{results_2011}
\alias{results_2011}
\title{Cerated data set to show off the package capabilities - new_df}
\format{A data frame some rows and some columns}
\usage{
results_2011
}
\description{
A manually created dataset showing the hypothetical scores of two divisions of students
}
\keyword{datasets}

