% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{results_2011}
\alias{results_2011}
\title{Data set created set to show off the package capabilities - Results of students for 2011}
\format{
A data frame 13 rows and 8 columns
}
\usage{
results_2011
}
\description{
A manually created dataset showing the hypothetical scores of two divisions of students
\itemize{
  \item Division The division to which the student belongs
  \item Student Name of the Student
  \item Maths, Physics, Chemistry, Art Scores of the student across different subjects
  \item Discipline, PE Grades of the students across different subjects
}
}
\keyword{datasets}
