% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnsOutputs.R
\name{view_html}
\alias{view_html}
\title{View Comparison output HTML}
\usage{
view_html(comparison_output)
}
\arguments{
\item{comparison_output}{output from the comparisonDF compare function}
}
\description{
Some versions of Rstudio doesn't automatically show the html pane for the html output. This is a workaround
}
